--- projects/standard-bridge/example.json
@@ -1,23 +1,0 @@
-{
-  "name": "Example Project Name",
-  "website": "https://example-project.io",
-  "twitter": "https://twitter.com/example-project",
-  "logoURI": "https://ethereum-optimism.github.io/logos/example.png",
-  "description": "A description of your project under 150 chars",
-  "tokens": [
-    {
-      "chainId": 42,
-      "address": "0x8e4d1588a1F3aa57875B3c8B328935513dB72f81",
-      "name": "Example Project",
-      "symbol": "EXMP",
-      "decimals": 18
-    },
-    {
-      "chainId": 69,
-      "address": "0x23aA4268E0CBb0F1c83DcA7c413b9f0C4a448A2B",
-      "name": "Example Project",
-      "symbol": "EXMP",
-      "decimals": 18
-    }
-  ]
-}
