+++ scripts/validateProjectInfo.ts
@@ -1,108 +1,37 @@
 import fs from "fs";
 import path from "path";
-import dotenv from "dotenv";
 import { providers, Contract } from "ethers";
 import tokenInterface from "../src/tokenInterface.json";
-import opTokenList from "../optimism.tokenlist.json";
-// TODO this needs to go
 import fetch from "node-fetch";
-import newTokenFile from "../projects/standard-bridge/exmpleProject/projectInfo.json";
-
+import {
+  infuraKey,
+  factoryDeployer,
+  network,
+  oppositeChainIdMap,
+  chainIdLayerMap,
+  validConfiguration,
+  validConfigurations,
+  standardBridgeAddress,
+  extensions,
+  networkURLMap,
+  networkMap,
+  TokenListing,
+  filePath,
+  readFile,
+  writeFile
+} from "./utils";
+import dotenv from "dotenv";
 dotenv.config();
 
 // FUTURE add flag to validate a single project passed in or all projects
 
-export const network = {
-  MAINNET: 1,
-  OP_MAINNET: 10,
-  KOVAN: 42,
-  OP_KOVAN: 69
-};
-
-export const oppositeChainIdMap = {
-  [network.MAINNET]: network.OP_MAINNET,
-  [network.KOVAN]: network.OP_KOVAN,
-  [network.OP_MAINNET]: network.MAINNET,
-  [network.OP_KOVAN]: network.KOVAN
-};
-
-export const chainIdLayerMap = {
-  [network.MAINNET]: 1,
-  [network.KOVAN]: 1,
-  [network.OP_MAINNET]: 2,
-  [network.OP_KOVAN]: 2
-};
-
-enum validConfiguration {
-  testing,
-  fullSuite,
-  unbridgeable
-}
-
-const validConfigurations = {
-  testing: [42, 69],
-  fullSuite: [1, 10, 42, 69],
-  unbridgeable: [10, 69]
-};
-
-const infuraKey = process.env.INFURA_KEY || "84842078b09946638c03157f83405213";
-
-const factoryDeployer:string = "0x4200000000000000000000000000000000000012";
-
-const standardBridgeAddress = {
-  1: "0x99C9fc46f92E8a1c0deC1b1747d010903E884bE1",
-  10: "0x4200000000000000000000000000000000000010",
-  42: "0x22F24361D548e5FaAfb36d1437839f080363982B",
-  69: "0x4200000000000000000000000000000000000010"
-}
-
-const networkURLMap = {
-  1: `https://mainnet.infura.io/v3/${infuraKey}`,
-  10: `https://mainnet.optimism.io`,
-  42: `https://kovan.infura.io/v3/${infuraKey}`,
-  69: `https://kovan.optimism.io`
-};
-
-const extensions = {
-  1: `{"optimismBridgeAddress":"${standardBridgeAddress[1]}"}`,
-  10: `{"optimismBridgeAddress":"${standardBridgeAddress[10]}"}`,
-  42: `{"optimismBridgeAddress":"${standardBridgeAddress[42]}"}`,
-  69: `{"optimismBridgeAddress":"${standardBridgeAddress[69]}"}`
-}
-
-const networkMap = {
-  1: "Mainnet",
-  10: "Optimistic Ethereum",
-  42: "Kovan",
-  69: "Optimistic Kovan"
-};
-
-type TokenListing = {
-  chainId: number;
-  address: string;
-  name: string;
-  symbol: string;
-  decimals: number;
-  logoURI: string;
-  factoryDeployed: Boolean;
-  extensions?: string;
-};
-
-let filePath = './projects/standard-bridge/';
-
 async function main() {
   let fileName = process.env.PROJECT_FILE_NAME;
-
-  filePath = './projects/standard-bridge/' + fileName + '/';
-  
-  let projectInfoPath = filePath + '/projectInfo.json';
+  let projectFilePath = filePath + fileName + '/';;
+  let projectInfoPath = projectFilePath + '/projectInfo.json';
   let resolvedPath = path.resolve(projectInfoPath);
   let receivedFile = readFile(resolvedPath);
-  // let parsedFile 
-
   let newToken = JSON.parse(receivedFile);
-
-  // // TODO verify off chain data
 
   let tokenChains:Array<number> = new Array;
   for(const token of newToken.tokens) {
@@ -337,29 +266,6 @@ function checkConfig(tokens: Array<number>): validConfiguration {
   }
 }
 
-function readFile(path: any): any {
-  try {
-    const data = fs.readFileSync(path, {encoding:'utf8', flag:'r'});
-    return data;
-  } catch (err) {
-    throw Error("Unable to read file.\n" + 
-      "Potential Fix: Please ensure your file is named correctly and located in the correct folder."
-    );
-  }
-}
-
-function writeFile(path: any, data: string) {
-  try {
-    fs.writeFileSync(path, data, {
-      flag: "w"
-    });
-  } catch (err) {
-    throw Error("Unable to write to file.\n" + 
-      "Potential Fix: Please ensure your file is named correctly and located in the correct folder."
-    );
-  }
-}
-
 async function getContractDeployer(
   address: string,
   chainId: number
@@ -381,6 +287,8 @@ async function getContractDeployer(
   );
   return deployerAddress.includes("GENESIS") ? "GENESIS" : deployerAddress;
 }
+
+// TODO verify off chain data
 
 main()
   .then(() => {
